backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
passlib[bcrypt]
python-multipart
httpx
python-dotenv
Pillow
numpy
tensorflow
pydantic[email]
